RakaBot/requirements.txt
python-telegram-bot
python-dotenv
transformers
torch
requests